+++ server/src/main/java/com/seb41_main_018/mainproject/content/controller/ContentController.java
@@ -1,4 +1,4 @@
-package com.seb41_main_018.mainproject.post.controller;
+package com.seb41_main_018.mainproject.content.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -6,20 +6,20 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @Validated
-@RequestMapping("/post")
-public class PostController {
+@RequestMapping("/content")
+public class ContentController {
 
     // 게시글 생성
     @PostMapping
-    public ResponseEntity postPost(
+    public ResponseEntity postContent(
             // TODO: DTO
     ) {
         return null;
     }
 
     // 게시글 단건 조회
-    @GetMapping("/{postId}")
-    public ResponseEntity getPost(@PathVariable("postId") Long postId
+    @GetMapping("/{contentId}")
+    public ResponseEntity getContent(@PathVariable("contentId") Long contentId
                                   // TODO: DTO
     ) {
         return null;
@@ -27,22 +27,22 @@ public class PostController {
 
     // 게시글 전체 조회
     @GetMapping
-    public ResponseEntity getPosts(
+    public ResponseEntity getContents(
     ) {
         return null;
     }
 
     // 게시글 수정
-    @PatchMapping("/{postId}")
-    public ResponseEntity patchPost(@PathVariable("postId") Long postId
+    @PatchMapping("/{contentId}")
+    public ResponseEntity patchContent(@PathVariable("contentId") Long contentId
                                     // TODO: DTO
     ) {
         return null;
     }
 
     // 게시글 삭제
-    @DeleteMapping("/{postId}")
-    public ResponseEntity deletePost(@PathVariable("postId") Long postId
+    @DeleteMapping("/{contentId}")
+    public ResponseEntity deleteContent(@PathVariable("contentId") Long contentId
                                      // TODO: DTO
     ) {
         return null;
